Standalone ebulletin route, no top-level about controller
