+++ controllers/ebulletin/views/ebulletin.njk
